Added test2.py
Test2.py runs using the Naomi plugins. I wanted a place to try
running Naomi with continuous streaming inputs and an action queue.
So far, I have done this by modifying the application.run() method
and the mic object. I think the logical place to put the main loop
is in the conversation.handleForever() method. This is where it
currently is, just uses the mic.listen() method to get audio input.
Instead, I need to make that a loop that just grabs audio from the
vad plugin and adds it to a queue, then starts a thread to process
audio from the queue.

Next I will be tackling the expect() method. This is a method that
should block while waiting for the next audio transcription, while
also using a custom vocabulary.